requirements.txt
influxdb3-python==0.3.6
paho-mqtt==2.0.0
python-dotenv==1.0.1